refactor functions from eggd_artemis.py to separate utils/ module for clarity
